requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiohttp==3.8.4
    # via
    #   aiohttp-client-cache
    #   discord-py
aiohttp-client-cache==0.8.1
    # via -r requirements.in
aiosignal==1.3.1
    # via aiohttp
aiosqlite==0.18.0
    # via -r requirements.in
async-timeout==4.0.2
    # via aiohttp
attrs==22.2.0
    # via
    #   aiohttp
    #   aiohttp-client-cache
charset-normalizer==3.1.0
    # via aiohttp
discord-py==2.2.2
    # via -r requirements.in
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via yarl
itsdangerous==2.1.2
    # via aiohttp-client-cache
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
python-dotenv==1.0.0
    # via -r requirements.in
python-forge==18.6.0
    # via aiohttp-client-cache
six==1.16.0
    # via url-normalize
url-normalize==1.4.3
    # via aiohttp-client-cache
yarl==1.8.2
    # via aiohttp
